% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status_updates.R
\name{get_status_updates}
\alias{get_status_updates}
\title{Populate workplanr database with data}
\usage{
get_status_updates(idb, token)
}
\arguments{
\item{staff}{Names of staff members}

\item{projects}{Names of projects}

\item{project_phases}{List of phases in any project in order of execution}

\item{project_roles}{List of project roles}

\item{out_of_office}{Names of staff that are going to be out of the office}

\item{public_holidays}{A data frame of dates of public holidays}

\item{roles_responsibilities}{Data frame for responsibilites of roles across project phases}

\item{time_estimates}{Time estimates of how long each phase will take in relation}

\item{staff_name_for_unassigned_work}{Dummy staff member to assign work to}
}
\description{
Populate workplanr database with data
}
